Best practices
Duplicating the common java files to grpc-client and grpc-service projects. This is done in order to follow the microservice best practice. The java files inside the grpc-common project were kind of forming a shared library between grpc-client and grpc-service. There are 2 main reasons for duplicating the code into each projects:
1) If grpc-common forms a shared library with common java code, we form a tight dependency between the grpc-common and all the microservices which are using it. Any change in the grpc-common will eventually call for a redeployment of other microservices. This violates the microservice concept of being able to build and deploy independently
2) The common code is written in Java. Tomorrow if there is another microservice which is written in some other language, that service will not be able to consume the shared library directly (or maybe it cannot consume it at all).